pyproject.toml
[project]
name = "bloomhub-be"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0",
    "psycopg[binary]>=3.1",
    "dj-database-url>=2.2",
    "python-dotenv>=1.0",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.8",
    "pytest-django>=4.8",
    "pre-commit>=3.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | migrations
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [".git", ".venv", "__pycache__", "migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
